+++ checkBeforeRun.md
@@ -1,0 +1,50 @@
+
+
+### commit message
+> **build** : 시스템 또는 외부 종속성에 영향을 미치는 변경사항 (npm, gulp, yarn 레벨) <br><br>
+**ci** : ci구성파일 및 스크립트 변경<br><br>
+**chore** : 패키지 매니저 설정할 경우, 코드 수정 없이 설정을 변경
+**docs** : documentation 변경<br><br>
+**feat** : 새로운 기능<br><br>
+**fix** : 버그 수정<br><br>
+**perf** : 성능 개선<br><br>
+**refactor** : 버그를 수정하거나 기능을 추가하지 않는 코드 변경, 리팩토링<br><br>
+**style** : 코드 의미에 영향을 주지 않는 변경사항 ( white space, formatting, colons )<br><br>
+**test** : 누락된 테스트 추가 또는 기존 테스트 수정<br><br>
+**revert** : 작업 되돌리기<br><br>
+
+
+
+### File structure
+<table>
+    <tr>
+        <td></td>
+        <td>back-end</td>
+        <td>front-end script</td>
+        <td>views</td>
+    </tr>
+    <tr>
+        <td>FOLDER</td>
+        <td>/</td>
+        <td>/public</td>
+        <td>/views</td>
+    </tr>
+    <tr>
+        <td>01</td>
+        <td>server.js</td>
+        <td> </td>
+        <td>studyList.ejs</td>
+    </tr>
+    <tr>
+        <td>02</td>
+        <td>server.js</td>
+        <td>chat_app.js</td>
+        <td>studyHome.ejs</td>
+    </tr>
+    <tr>
+        <td>03</td>
+        <td>server.js</td>
+        <td>script.js</td>
+        <td>videoChat.ejs</td>
+    </tr>
+</table>
