+++ public/chat_app.js
@@ -1,13 +1,11 @@
 const socket = io();
 
-const welcome = document.getElementById('welcome');
-const form = welcome.querySelector('form');
-const room = document.getElementById('room');
+const room = document.getElementById('chatRoom');
 
-room.hidden = true;
+let userName = NICKNAME;
+let roomName = ROOM_ID;
 
-let roomName;
-
+//functions
 function addMessage(message) {
   const ul = room.querySelector('ul');
   const li = document.createElement('li');
@@ -32,12 +30,6 @@ function handleMessageSubmit(event) {
   input.value = '';
 }
 
-function handleNicknameSubmit(event) {
-  event.preventDefault();
-  const input = room.querySelector('#name input');
-  socket.emit('nickname', input.value);
-}
-
 function handleNoticeSubmit(event) {
   event.preventDefault();
   const input = room.querySelector('#notice input');
@@ -48,36 +40,18 @@ function handleNoticeSubmit(event) {
   input.value = '';
 }
 
-function showRoom() {
-  welcome.hidden = true;
-  room.hidden = false;
-  const h3 = room.querySelector('h3');
-  h3.innerText = `Room ${roomName}`;
+function startChat() {
+  const h1 = document.querySelector("h1");
+  h1.innerText = `TEAM12 - STUDY HOME : ${roomName}`;
   const msgForm = room.querySelector('#msg');
-  const nameForm = room.querySelector('#name');
   const noticeForm = room.querySelector('#notice');
   msgForm.addEventListener('submit', handleMessageSubmit);
-  nameForm.addEventListener('submit', handleNicknameSubmit);
   noticeForm.addEventListener('submit', handleNoticeSubmit);
 }
 
-function handleRoomSubmit(event) {
-  event.preventDefault();
-  const input = form.querySelector('input');
-  socket.emit('enter_room', input.value, showRoom);
-  roomName = input.value;
-  input.value = '';
-}
-
-form.addEventListener('submit', handleRoomSubmit);
-
-socket.on('welcome', (user) => {
-  addMessage(`${user} arrived!`);
-});
-
-socket.on('bye', (left) => {
-  addMessage(`${left} left ㅠㅠ`);
-});
+// socket
+socket.emit('enter_chat_room', roomName, startChat);
+socket.emit('nickname', userName);
 
 socket.on('new_message', addMessage);
 socket.on('new_notice', addNotice);
