+++ server.js
@@ -18,7 +18,12 @@ app.use(express.urlencoded({extended:false}));      // body-parser
 app.use('/peerjs', peerServer);
 
 app.get("/", (req, res) => {
-    res.render("studyHome");
+    res.render("studyList");
+});
+app.post("/studyHome", (req, res) => {
+    studyRoomId = req.body.StudyName;
+    Nickname = req.body.Nickname;
+    res.render("studyHome", { roomID: studyRoomId, nickname: Nickname });
 });
 app.post("/videoChat", (req, res) => {
     studyRoomId = req.body.StudyName;
@@ -28,8 +33,32 @@ app.post("/videoChat", (req, res) => {
 
 
 io.on("connection", socket => {
-
-
+    //msg chat
+    socket['nickname'] = 'Anonymous';
+    // socket.onAny((event) => {
+    //     console.log(`Socket Event: ${event}`);
+    // });
+    socket.on('enter_chat_room', (chatRoomName, done) => {
+        socket.join(chatRoomName);
+        done();
+        //socket.to(roomName).emit('welcome', socket.nickname);
+    });
+    // socket.on('disconnecting', () => {
+    //     socket.rooms.forEach((room) =>
+    //     socket.to(room).emit('bye', socket.nickname)
+    //     );
+    // });
+    socket.on('new_message', (msg, room, done) => {
+        socket.to(room).emit('new_message', `${socket.nickname}: ${msg}`);
+        done(); //triggers function located at frontend
+    });
+    socket.on('nickname', (nickname) => (socket['nickname'] = nickname));
+    socket.on('new_notice', (msg, room, done) => {
+        socket.to(room).emit('new_notice', `NOTICE: ${msg}`);
+        done();
+    });
+    
+    //video chat
     socket.on("join-room", (roomId, userId) => {
         socket.join(roomId);
         // version A
@@ -47,5 +76,5 @@ io.on("connection", socket => {
         });
     })
 });
-
-server.listen(process.env.PORT||3000);
+const handleListen = () => console.log(`Listening on http://localhost:3000`);
+server.listen(process.env.PORT||3000, handleListen);
