--- views/chat_home.pug
@@ -1,30 +1,0 @@
-doctype html
-html(lang="en")
-    head
-        meta(charset="UTF-8")
-        meta(http-equiv="X-UA-Compatible", content="IE=edge")
-        meta(name="viewport", content="width=device-width, initial-scale=1.0")
-        title Noom
-        link(rel="stylesheet", href="https://unpkg.com/mvp.css")
-    body
-        header
-            h1 Noom
-        main
-            div#welcome
-                form
-                    input(placeholder="room name", required, type="text")
-                    button Enter Room
-            div#room
-                h3
-                ul
-                form#name
-                    input(placeholder="nickname", required, type="text")
-                    button Save
-                form#notice
-                    input(placeholder="notice", type="text")
-                    button Send
-                form#msg
-                    input(placeholder="message", required, type="text")
-                    button Send
-        script(src="/socket.io/socket.io.js") 
-        script(src="/public/js/app.js")
