+++ views/studyList.ejs
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    
+    <title>STUDY LIST</title>
+    <link rel="stylesheet" type="text/css" href="https://unpkg.com/mvp.css@1.12/mvp.css"></style>
+</head>
+<body>
+    <header>
+        <h1> TEAM12 - STUDY LIST</h1>
+    </header>
+
+    <main>
+        <div>
+            <form action="/studyHome" method="post">
+                <input type="text" id="Nickname" name="Nickname" placeholder="Your Name">
+                <input type="text" id="StudyName" name="StudyName" placeholder="Your Study">
+                <input type="submit" id="btnEnterRoom" value="Enter Room">
+            </form>
+        </div>
+    </main>
+</body>
+</html>
