+++ views/videoChat.ejs
@@ -9,6 +9,8 @@
     <script>
         const ROOM_ID = "<%= roomID %>"
         const NICKNAME = "<%= nickname %>"
+
+        console.log(ROOM_ID, NICKNAME);
     </script>
     <script defer src="https://unpkg.com/peerjs@1.4.7/dist/peerjs.min.js"></script>
     <script defer src="/socket.io/socket.io.js"></script>
